Allow ConnectionFields to have ObjectTypes as per relay spec
